reafctor: on auth state changed
